requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
colorama==0.4.6
